change extraction of optional properties
