+++ ui/plugins/eu.esdihumboldt.hale.ui.views.schemas/src/eu/esdihumboldt/hale/ui/views/schemas/explorer/EntitySchemaExplorer.java
@@ -29,8 +29,8 @@ import eu.esdihumboldt.hale.common.align.model.EntityDefinition;
 import eu.esdihumboldt.hale.common.schema.SchemaSpaceID;
 import eu.esdihumboldt.hale.common.schema.model.Definition;
 import eu.esdihumboldt.hale.ui.common.definition.viewer.InheritedPropertiesFilter;
-import eu.esdihumboldt.hale.ui.common.definition.viewer.OptionalPropertiesFilter;
 import eu.esdihumboldt.hale.ui.service.entity.EntityDefinitionService;
+import eu.esdihumboldt.hale.ui.service.entity.OptionalPropertiesFilter;
 import eu.esdihumboldt.hale.ui.service.entity.util.ContentProviderAction;
 import eu.esdihumboldt.hale.ui.service.entity.util.EntityTypeIndexContentProvider;
 import eu.esdihumboldt.hale.ui.service.entity.util.EntityTypeIndexHierarchy;
